Remove %llu and long long from AMUDP
* Tweak temporary types in AMUDP_DumpStatistics()

(cherry picked from commit 0057d1c051dda60d7da0f46e7c51bdc2980343e1)

+++ other/amudp/amudp_reqrep.cpp
@@ -61,7 +61,7 @@ extern void AMUDP_InitRetryCache() {
   #if 0  // for debugging retry calc
     for (int i=0; i < STATIC_RETRIES*2; i++) {
       amudp_cputick_t tick = REQUEST_TIMEOUT_TICKS(i);
-      printf("Timeout %2i: %9llu us, %9llu ticks\n",i,(unsigned long long)tick/us2ticks(1),(unsigned long long)tick);
+      printf("Timeout %2i: %9" PRIu64 " us, %9" PRIu64 " ticks\n",i,tick/us2ticks(1),tick);
     }
   #endif
 }
